pyproject.toml
[tool.poetry]
name = "powershell-chatbot"
version = "0.1.0"
description = "Powershell interface for GPT4-turbo chatbot assistant"
authors = ["James O'Reilly <joreilly86@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "1"
pyperclip = "^1.8.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
